Fix an error
