+++ Week 1 assignment/Aayush/Q1.cpp
@@ -1,0 +1,17 @@
+#include<iostream>
+using namespace std;
+int main()
+{
+    for(int i=1;i<=100;i++)
+    {
+        if(i%3==0 && i%5==0)
+            cout<<"FizzBuzz"<<endl;
+        else if(i%3==0)
+            cout<<"Fizz"<<endl;
+        else if(i%5==0)
+            cout<<"Buzz"<<endl;
+        else
+            cout<<i<<endl;
+    }
+    return 0;
+}
