+++ Week 1 assignment/Aayush/Q1.exe
