+++ Week 1 assignment/Aayush/Q2.exe
