+++ Week 1 assignment/Aayush/Q3.exe
