+++ Week 1 assignment/Aayush/Q4.exe
