+++ Week 1 assignment/Aayush/Q5.exe
