+++ Week 1 assignment/Aayush/Q6.cpp
@@ -1,0 +1,19 @@
+#include<iostream>
+using namespace std;
+
+int main() {
+    int n;
+    cout << "Enter a number" << endl;   
+    cin >> n;
+
+    for (int i = 1; i <= n; i++) {
+        for (int j = 1; j <= i; j++) {
+            for (int k = 1; k <= i; k++) {
+                cout << j;
+            }
+        }
+        cout << endl;
+    }
+
+    return 0;
+}
