+++ Week 1 assignment/Aayush/Q6.exe
