+++ Week 1 assignment/Aayush/Q7.cpp
@@ -1,0 +1,23 @@
+#include <iostream>
+using namespace std;
+
+int main()
+{
+    int n;
+    cout << "Enter a number: ";
+    cin >> n;
+
+    int num = 1; 
+    for (int i = 1; i <= n; i++) {
+        for (int j = 1; j <= n; j++) {
+            cout << num;
+            if (j < n) {
+                cout << "*";
+            }
+            num++;
+        }
+        cout << endl; 
+    }
+
+    return 0;
+}
