+++ Week 1 assignment/Aayush/Q7.exe
